ai_backend/requirements.txt
fastapi
uvicorn
pillow
torch
yolov3 @ git+https://github.com/ultralytics/yolov3.git
